refactor(inspections-api): extract media sub-router
Photo/media/video upload, serve, crop, reorder, attach, annotations, and
cover-photo concierge approve (19 routes) move to server/api/inspections/media.ts.
Handler bodies byte-identical; external paths and OpenAPI/RPC surface unchanged.
